Fix 64bit cast issue
git-svn-id: svn+ssh://svn.gna.org/svn/xenomai/trunk@2638 c6d672ea-8702-0410-b560-f74c916a59fe

+++ src/testsuite/clocktest/clocktest.c
@@ -207,7 +207,7 @@ int main(int argc, char *argv[])
     for (i = 0; i < cpus; i++) {
         per_cpu_data[i].first_round = 1;
         pthread_create(&per_cpu_data[i].thread, NULL, cpu_thread,
-                    (void *)i);
+		       (void *)(long)i);
     }
 
     printf("== Tested clock: %d (", clock_id);
